refactor: 테스트 클래스의 위치를 프로덕션 클래스와 동일하게 수정
