+++ src/test/java/roomescape/controller/roomescape/admin/ThemeControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.roomescape;
+package roomescape.controller.roomescape.admin;
 
 import static org.hamcrest.Matchers.is;
